fixed and cleaned code
