+++ src/app/view/viewManager.js
@@ -893,6 +893,54 @@ export default class ViewManager {
     }
 
     /**
+     * Center all the components on the canvas
+     * @returns {void}
+     * @memberof ViewManager
+     */
+    centerAll() {
+        let components = Registry.currentDevice.getComponents();
+        let connections = Registry.currentDevice.getConnections();
+
+        if (components.length > 0) {
+            // Find the center of the desige
+            let leftBound, rightBound, upperBound, lowerBound;
+            [leftBound, upperBound] = components[0].getParams().getValue("position");
+            [rightBound, lowerBound] = components[0].getParams().getValue("position");
+            components.forEach(component => {
+                let position = component.getParams().getValue("position");
+                leftBound = Math.min(leftBound, position[0]);
+                rightBound = Math.max(rightBound, position[0]);
+                upperBound = Math.min(upperBound, position[1]);
+                lowerBound = Math.max(lowerBound, position[1]);
+            });
+            connections.forEach(connection => {
+                let position = connection.getParams().getValue("start");
+                leftBound = Math.min(leftBound, position[0]);
+                rightBound = Math.max(rightBound, position[0]);
+                upperBound = Math.min(upperBound, position[1]);
+                lowerBound = Math.max(lowerBound, position[1]);
+            });
+            let centerX = (leftBound + rightBound) / 2;
+            let centerY = (upperBound + lowerBound) / 2;
+
+            // Calculate the change of position
+            let changeX = Registry.currentDevice.getXSpan() / 2 - centerX;
+            let changeY = Registry.currentDevice.getYSpan() / 2 - centerY;
+
+            // Move every component to its center
+            components.forEach(component => {
+                let position = component.getParams().getValue("position");
+                component.updateComponetPosition([position[0] + changeX, position[1] + changeY]);
+            });
+
+            // Move every connection to its center
+            connections.forEach(connection => {
+                connection.updateConnectionPosition(changeX, changeY);
+            });
+        }
+    }
+
+    /**
      * Updates the default feature parameter
      * @param {string} typeString
      * @param {string} setString
